+++ Assets/Plugins/BlackSmith.Domain/Domain/Character/Player/PlayerCommonReconstructCommand.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 147c4628cf6f26b4bbf463d19aa9ca62
