+++ Assets/Plugins/BlackSmith.Domain/Tests/Domain/Character.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1dfacf49dda73f6409a18ff9f9509c07
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
