+++ Assets/Plugins/BlackSmith.Domain/Tests/Domain/Character/Battle.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b4d501ca6e923104da83a2fedda194c8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
