+++ Assets/Plugins/BlackSmith.Domain/Tests/Domain/Character/Battle/HealthPointTest.cs
