+++ Assets/Plugins/BlackSmith.Domain/Tests/Domain/Character/Player.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 13c11a26a179f8b4db9b40dd66ae3a58
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
