+++ Assets/Plugins/BlackSmith.Domain/Tests/Domain/Character/Player/ExperienceTest.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 640520a0296b2984084873ddb4ad8b93
