+++ Assets/Plugins/BlackSmith.Domain/Tests/Domain/Character/Player/PlayerCommonReconstructCommandTest.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 6df7fbf44930ecc4f9977a3e2b33c657
