+++ Assets/Plugins/BlackSmith.Domain/Tests/Domain/Character/Player/PlayerNameTest.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 36da9b1b393da57409c8ac9cb7c0dccf
