--- Assets/Plugins/BlackSmith.Domain/Tests/Domain/ExperienceTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8c84f7cabdcdd504487fffd6931a1f05
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
