+++ Assets/Plugins/BlackSmith.Domain/Tests/Domain/Player.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 615b7153babc97447a586c5a378dff93
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
